// frontend/types/types.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ClientMessage = { "type": "request_grid_state" } | { "type": "tile_update", col: number, row: number, data: HexData, };

export type HexData = { terrain: TerrainType, };

export type ServerMessage = { "type": "grid_state", width: number, height: number, tiles: Array<TileState>, } | { "type": "tile_update", col: number, row: number, data: HexData, };

export type TerrainType = "Wild" | "Mine" | "Turret" | "Slime";

export type TileState = { col: number, row: number, data: HexData, };